+++ wmflib/actions.py
@@ -11,11 +11,10 @@ class Actions:
     """Class to keep track and log a set of actions performed and their result with a nice string representation."""
 
     def __init__(self, name: Hashable):
-        """Initialize the instance.
+        """The instance gets initialized with the given name, that can represent a host or any other identifier.
 
-        When converted to string returns a nicely formatted representation of the instance and all its actions.
-
-        It exposes the following properties:
+        It allows to save actions performed on the given name with their status (success, warning, failure).
+        It exposes the following public properties:
 
         - ``name``: the name passed to the instance at instantiation time.
 
@@ -24,6 +23,30 @@ class Actions:
 
         - ``has_failures``: a :py:class:`bool` that is :py:data:`True` when at least one failed action was registered,
           :py:data:`False` otherwise.
+
+        When converted to a string it returns a nicely formatted representation of the instance and all its actions
+        that can be used as-is in Phabricator::
+
+            <name> (**<status>**)
+              - <action1>
+              - <action2>
+              - ...
+
+        The status is enclosed in double asterisks to be rendered as bold in Phabricator.
+
+        Examples:
+            ::
+
+                actions = Actions('host1001')
+                actions.success('Downtimed on Icinga')
+                actions.success('Restarted ntp')
+                print(actions)
+
+            The above code will print::
+
+                host1001 (**PASS**)
+                  - Downtimed on Icinga
+                  - Restarted ntp
 
         Arguments:
             name (typing.Hashable): the name of the set of actions to be registered.
@@ -41,15 +64,18 @@ class Actions:
             str: the string representation.
 
         """
-        actions = '\n'.join('  - {action}'.format(action=action) for action in self.actions)
-        return '{name} (**{status}**)\n{actions}'.format(name=self.name, status=self.status, actions=actions)
+        actions = '\n'.join(f'  - {action}' for action in self.actions)
+        return f'{self.name} (**{self.status}**)\n{actions}'
 
     @property
     def status(self) -> str:
         """Return the current status of the actions based on the worst result recorded.
 
         Returns:
-            str: the short string representation of the status, one of: ``PASS``, ``WARN``, ``FAIL``.
+            str:
+                * ``FAIL``: if there was at least one failure action registered
+                * ``WARN`` if there was at least one warning action and no ``FAIL`` actions registered
+                * ``PASS`` if only success actions were registered
 
         """
         if self.has_failures:
@@ -60,7 +86,7 @@ class Actions:
         return 'PASS'
 
     def success(self, message: str) -> None:
-        """Register a successful action.
+        """Register a successful action, it gets also logged with info level.
 
         Arguments:
             message (str): the action description.
@@ -69,7 +95,7 @@ class Actions:
         self._action(logging.INFO, message)
 
     def failure(self, message: str) -> None:
-        """Register a failed action.
+        """Register a failed action, it gets also logged with error level.
 
         Arguments:
             message (str): the action description.
@@ -79,7 +105,7 @@ class Actions:
         self.has_failures = True
 
     def warning(self, message: str) -> None:
-        """Register an action that require some attention.
+        """Register an action that require some attention, it gets also logged with warning level.
 
         Arguments:
             message (str): the action description.
@@ -109,7 +135,29 @@ class ActionsDict(dict):
     When accessing a missing key, the key itself is passed to the new :py:class:`wmflib.actions.Actions` instance
     as ``name``.
 
-    When converted to string returns a nicely formatted representation of the instance and all its items.
+    When converted to string returns a nicely formatted representation of the instance and all its items, that can be
+    used as-is in Phabricator.
+
+    Examples:
+        ::
+
+            actions = ActionsDict()
+            actions['host1001'].success('Downtimed on Icinga')
+            actions['host1001'].failure('**Failed to restart ntp**')  # Will be rendered in bold in Phabricator
+            actions['host2001'].warning('//Host with alerts on Icinga//')  # Will be rendered in italic in Phabricator
+            print(actions)
+
+        The above code will print::
+
+            - host1001 (**FAIL**)
+              - Downtimed on Icinga
+              - **Failed to restart ntp**
+
+            - host2001 (**WARN**)
+              - //Host not optimal on Icinga//
+
+        It will also include a final newline at the end of the block that doesn't get rendered in this documentation.
+
     """
 
     def __missing__(self, key: Hashable) -> Actions:
@@ -125,7 +173,7 @@ class ActionsDict(dict):
         """Custom string representation of the instance.
 
         Returns:
-            str: the string representation.
+            str: the string representation of each item in the dictionary, newline-separated.
 
         """
-        return '\n'.join('- {actions}\n'.format(actions=value) for key, value in self.items())
+        return '\n'.join(f'- {value}\n' for value in self.values())