+++ wmflib/decorators.py
@@ -16,7 +16,14 @@ logger = logging.getLogger(__name__)
 # TODO: use TypedDict once Python 3.7 support is removed
 @dataclass
 class RetryParams:
-    """Retry decorator parameters object."""
+    """Retry decorator parameters class.
+
+    This class represents the parameters that the :py:func:`wmflib.decorators.retry` decorator accepts and allow
+    to validate them.
+
+    It is also used for the `dynamic_params_callbacks` parameter of the same retry decorator to allow the callback to
+    safely modify the decorator's parameters at runtime.
+    """
 
     tries: int
     delay: timedelta
@@ -25,21 +32,20 @@ class RetryParams:
     failure_message: str
 
     def validate(self) -> None:
-        """Validate the consistency of the current field values.
+        """Validate the consistency of the current values of the instance properties.
 
         Raises:
             wmflib.exceptions.WmflibError: if any field has an invalid value.
 
         """
         if self.backoff_mode not in ('constant', 'linear', 'power', 'exponential'):
-            raise WmflibError('Invalid backoff_mode: {mode}'.format(mode=self.backoff_mode))
+            raise WmflibError(f'Invalid backoff_mode: {self.backoff_mode}')
 
         if self.backoff_mode == 'exponential' and self.delay.total_seconds() < 1:
-            raise WmflibError('Delay must be greater than 1 if backoff_mode is exponential, got {delay}'.format(
-                delay=self.delay))
+            raise WmflibError(f'Delay must be greater than 1 if backoff_mode is exponential, got {self.delay}')
 
         if self.tries < 1:
-            raise WmflibError('Tries must be a positive integer, got {tries}'.format(tries=self.tries))
+            raise WmflibError(f'Tries must be a positive integer, got {self.tries}')
 
         if not self.failure_message:
             raise WmflibError('A failure_message must be set.')
@@ -65,7 +71,7 @@ def ensure_wrap(func: Callable) -> Callable:
 
 
 @ensure_wrap
-def retry(  # pylint: disable=too-many-arguments
+def retry(
     func: Callable,
     *,
     tries: int = 3,
@@ -80,6 +86,51 @@ def retry(  # pylint: disable=too-many-arguments
     Note:
         The decorated function or method must be idempotent to avoid unwanted side effects.
         It can be called with or without arguments, in the latter case all the default values will be used.
+
+    Examples:
+        Define a function that polls for the existence of a file, retrying with the default parameters::
+
+            >>> from pathlib import Path
+            >>> from wmflib.decorators import retry
+            >>> from wmflib.exceptions import WmflibError
+            >>> @retry
+            ... def poll_file(path: Path):
+            ...     if not path.exists():
+            ...         raise WmflibError(f'File {path} not found')
+            ...
+            >>> poll_file(Path('/tmp'))
+            >>> poll_file(Path('/tmp/nonexistent'))
+            [1/3, retrying in 3.00s] Attempt to run '__main__.poll_file' raised: File /tmp/nonexistent not found
+            [2/3, retrying in 9.00s] Attempt to run '__main__.poll_file' raised: File /tmp/nonexistent not found
+            Traceback (most recent call last):
+              File "<stdin>", line 1, in <module>
+              File "/usr/lib/python3/dist-packages/wmflib/decorators.py", line 160, in wrapper
+                return func(*args, **kwargs)
+              File "<stdin>", line 4, in poll_file
+            wmflib.exceptions.WmflibError: File /tmp/nonexistent not found
+            >>>
+
+        Same example, but customizing the decorator parameters::
+
+            >>> from datetime import timedelta
+            >>> from pathlib import Path
+            >>> from wmflib.decorators import retry
+            >>> @retry(tries=5, delay=timedelta(seconds=30), backoff_mode='constant', failure_message='File not found',
+            ...        exceptions=(RuntimeError,))
+            ... def poll_file(path: Path):
+            ...     if not path.exists():
+            ...         raise RuntimeError(path)
+            ...
+            [1/5, retrying in 30.00s] File not found: /tmp/nonexistent
+            [2/5, retrying in 30.00s] File not found: /tmp/nonexistent
+            [3/5, retrying in 30.00s] File not found: /tmp/nonexistent
+            [4/5, retrying in 30.00s] File not found: /tmp/nonexistent
+            Traceback (most recent call last):
+              File "<stdin>", line 1, in <module>
+              File "/usr/lib/python3/dist-packages/wmflib/decorators.py", line 160, in wrapper
+                return func(*args, **kwargs)
+              File "<stdin>", line 5, in poll_file
+            RuntimeError: /tmp/nonexistent
 
     Arguments:
         func (function, method): the decorated function.
@@ -126,8 +177,7 @@ def retry(  # pylint: disable=too-many-arguments
 
     """
     if not failure_message:
-        failure_message = "Attempt to run '{module}.{qualname}' raised".format(
-            module=func.__module__, qualname=func.__qualname__)
+        failure_message = f"Attempt to run '{func.__module__}.{func.__qualname__}' raised"
 
     static_params: Dict[str, Any] = {
         'tries': tries,
@@ -179,10 +229,10 @@ def _exception_message(exception: BaseException) -> str:
         # reverse order from the built-in handler (i.e. newest exception first) since we aren't following a
         # traceback.
         if exception.__cause__ is not None:
-            message_parts.append('Caused by: {chained_exc}'.format(chained_exc=exception.__cause__))
+            message_parts.append(f'Caused by: {exception.__cause__}')
             exception = exception.__cause__
         else:  # e.__context__ is not None, due to the while condition.
-            message_parts.append('Raised while handling: {chained_exc}'.format(chained_exc=exception.__context__))
+            message_parts.append(f'Raised while handling: {exception.__context__}')
             exception = cast(BaseException, exception.__context__)  # Casting away the Optional.
     return '\n'.join(message_parts)
 
@@ -208,6 +258,6 @@ def get_backoff_sleep(backoff_mode: str, base: Union[int, float], index: int) ->
     elif backoff_mode == 'exponential':
         sleep = base ** index
     else:
-        raise ValueError('Invalid backoff_mode: {mode}'.format(mode=backoff_mode))
+        raise ValueError(f'Invalid backoff_mode: {backoff_mode}')
 
     return sleep