+++ wmflib/fileio.py
@@ -4,7 +4,7 @@ import logging
 
 from contextlib import contextmanager
 from datetime import timedelta
-from pathlib import Path
+from os import PathLike
 from typing import Generator, IO
 
 from wmflib.decorators import retry
@@ -23,11 +23,21 @@ class LockError(FileIOError):
 
 
 @contextmanager
-def locked_open(file_path: Path, file_mode: str = 'r', *, timeout: int = 10) -> Generator[IO, None, None]:
+def locked_open(file_path: PathLike, file_mode: str = 'r', *, timeout: int = 10) -> Generator[IO, None, None]:
     """Context manager to open a file with an exclusive lock on it and a retry logic.
 
+    Examples:
+        ::
+
+            from wmflib.fileio import locked_open
+            with locked_open('existing.file') as f:
+                text = f.read()
+
+            with locked_open('new.out', 'w') as f:
+                f.write('Some text')
+
     Arguments:
-        file_path (pathlib.Path): the file path to open.
+        file_path (os.PathLike): the file path to open.
         file_mode (str, optional): the mode in which the file is opened, see :py:func:`open` for details.
         timeout (int, optional): the total timeout in seconds to wait to acquire the exclusive lock before giving up.
             Ten tries will be attempted to acquire the lock within the timeout.
@@ -40,7 +50,7 @@ def locked_open(file_path: Path, file_mode: str = 'r', *, timeout: int = 10) -> 
 
     """
     tries = 10
-    with open(file_path, file_mode) as fd:
+    with open(file_path, file_mode, encoding='utf-8') as fd:
         try:
             # Decorate the call to the locking function to retry acquiring the lock:
             # decorator(decorator_args)(function)(function_args)
@@ -53,7 +63,7 @@ def locked_open(file_path: Path, file_mode: str = 'r', *, timeout: int = 10) -> 
             )(fcntl.flock)(fd, fcntl.LOCK_EX | fcntl.LOCK_NB)
             logger.debug('Acquired exclusive lock on %s', file_path)
         except OSError as e:
-            raise LockError('Unable to acquire exclusive lock on {name}'.format(name=file_path)) from e
+            raise LockError(f'Unable to acquire exclusive lock on {file_path}') from e
 
         try:
             yield fd