+++ wmflib/interactive.py
@@ -24,6 +24,18 @@ class AbortError(WmflibError):
 def ask_input(message: str, choices: Sequence[str]) -> str:
     """Ask the user for input in interactive mode.
 
+    Examples:
+        ::
+
+            >>> choices = ['A', 'B']
+            >>> response = ask_input(f'Choose a door between {choices}', choices)
+            ==> Choose a door between ['A', 'B']
+            > a
+            ==> Invalid response, please type one of: A,B. After 3 wrong answers the task will be aborted.
+            > A
+            >>> response
+            'A'
+
     Arguments:
         message (str): the message to be printed before asking for confirmation.
         choices (sequence): the available choices of possible answers that the user can give. Values must be strings.
@@ -35,25 +47,42 @@ def ask_input(message: str, choices: Sequence[str]) -> str:
         wmflib.interactive.InputError: if not in a TTY or on too many invalid answers.
 
     """
-    prefix = '\x1b[36m>>>\x1b[39m'  # Cyan >>> prefix
+    prefix = '\x1b[36m==>\x1b[39m'  # Cyan ==> prefix
     if not sys.stdout.isatty():
         raise InputError('Not in a TTY, unable to ask for input')
 
-    print('{prefix} {message}'.format(prefix=prefix, message=message))
+    print(f'{prefix} {message}')
 
     for _ in range(3):
         response = input('> ')
         if response in choices:
             return response
 
-        print('{prefix} Invalid response, please type one of: {choices}. '
-              'After 3 wrong answers the task will be aborted.'.format(prefix=prefix, choices=','.join(choices)))
+        print(f'{prefix} Invalid response, please type one of: {",".join(choices)}. '
+              'After 3 wrong answers the task will be aborted.')
 
     raise InputError('Too many invalid answers')
 
 
 def ask_confirmation(message: str) -> None:
     """Ask the use for confirmation in interactive mode.
+
+    Examples:
+        ::
+
+            >>> ask_confirmation('Ready to continue?')
+            ==> Ready to continue?
+            Type "go" to proceed or "abort" to interrupt the execution
+            > go
+            >>> ask_confirmation('Ready to continue?')
+            ==> Ready to continue?
+            Type "go" to proceed or "abort" to interrupt the execution
+            > abort
+            Traceback (most recent call last):
+              File "<stdin>", line 1, in <module>
+              File "/usr/lib/python3/dist-packages/wmflib/interactive.py", line 69, in ask_confirmation
+                raise AbortError('Confirmation manually aborted')
+            wmflib.interactive.AbortError: Confirmation manually aborted
 
     Arguments:
         message (str): the message to be printed before asking for confirmation.
@@ -71,6 +100,25 @@ def ask_confirmation(message: str) -> None:
 
 def confirm_on_failure(func: Callable, *args: Any, **kwargs: Any) -> Any:
     """Execute a function asking for confirmation to retry, abort or skip.
+
+    Examples:
+        ::
+
+            >>> def test(fail=False):
+            ...     if fail:
+            ...         raise RuntimeError('Failed')
+            ...
+            >>> confirm_on_failure(test)
+            >>> confirm_on_failure(test, fail=True)
+            Failed to run __main__.test: Failed
+            ==> What do you want to do? "retry" the last command, manually fix the issue and "skip" the last command to
+                continue the execution or completely "abort" the execution.
+            > retry
+            Failed to run __main__.test: Failed
+            ==> What do you want to do? "retry" the last command, manually fix the issue and "skip" the last command to
+                continue the execution or completely "abort" the execution.
+            > skip
+            >>>
 
     Arguments:
         func (callable): the function/method to execute.
@@ -101,13 +149,19 @@ def confirm_on_failure(func: Callable, *args: Any, **kwargs: Any) -> Any:
             if response == 'skip':
                 return None
             if response == 'abort':
-                raise AbortError('Task manually aborted')
+                raise AbortError('Task manually aborted') from e
         else:
             return ret
 
 
 def get_username() -> str:
     """Detect and return the name of the effective running user even if run as root.
+
+    Examples:
+        ::
+
+            >>> get_username()
+            'user'
 
     Returns:
         str: the name of the effective running user or ``-`` if unable to detect it.
@@ -128,6 +182,12 @@ def get_username() -> str:
 def ensure_shell_is_durable() -> None:
     """Ensure it is running either in non-interactive mode or in a screen/tmux session, raise otherwise.
 
+    Examples:
+        ::
+
+            >>> ensure_shell_is_durable()  # Will raise if not in a tmux/screen session
+            >>>
+
     Raises:
         wmflib.exceptions.WmflibError: if in a non-durable shell session.
 
@@ -142,6 +202,17 @@ def ensure_shell_is_durable() -> None:
 def get_secret(title: str, *, confirm: bool = False) -> str:
     """Ask the user for a secret e.g. password.
 
+    Examples:
+        ::
+
+            >>> secret = get_secret('Secret key')
+            Secret key:
+            Secret must be at least 6 characters. try again:
+            >>> secret = get_secret('Secret key', confirm=True)  # Will raise if the confirmation doesn't match
+            Secret key:
+            Again, just to be sure:
+            >>>
+
     Arguments:
         title (str): The message to show the user.
         confirm (bool, optional): If :py:data:`True` ask the user to confirm the password.
@@ -149,14 +220,17 @@ def get_secret(title: str, *, confirm: bool = False) -> str:
     Returns:
         str: the secret.
 
+    Raises:
+        wmflib.exceptions.WmflibError: if the password confirmation does not match and confirm is :py:data:`True`.
+
     """
-    new_secret = getpass.getpass(prompt='{}: '.format(title))
+    new_secret = getpass.getpass(prompt=f'{title}: ')
 
     while len(new_secret) < MIN_SECRET_SIZE:
         new_secret = getpass.getpass(
-            prompt='Secret must be at least {} characters. try again: '.format(MIN_SECRET_SIZE))
+            prompt=f'Secret must be at least {MIN_SECRET_SIZE} characters. try again: ')
 
     if confirm and new_secret != getpass.getpass(prompt='Again, just to be sure: '):
-        raise WmflibError('{}: Passwords did not match'.format(title))
+        raise WmflibError(f'{title}: Passwords did not match')
 
     return new_secret