+++ wmflib/tests/sphinx_checker.py
@@ -6,15 +6,16 @@
   Sphinx would raise a warning but not fail in this case.
 """
 import argparse
-import os
 import pkgutil
 import sys
+
+from pathlib import Path
 
 import wmflib
 
 
-DOC_API_BASE_PATH = 'doc/source/api'
-DOC_API_INDEX_PATH = os.path.join(DOC_API_BASE_PATH, 'index.rst')
+DOC_API_BASE_PATH = Path('doc/source/api')
+DOC_API_INDEX_PATH = DOC_API_BASE_PATH / 'index.rst'
 API_INDEX_PREFIX = '    wmflib.'
 EXCLUDED_NAMES = ()
 
@@ -24,8 +25,8 @@ def main(base_path):
     wmflib_modules = {name for _, name, ispkg in pkgutil.iter_modules(wmflib.__path__)
                       if not ispkg and name not in EXCLUDED_NAMES}
 
-    doc_path = os.path.join(base_path, DOC_API_INDEX_PATH)
-    with open(doc_path) as f:
+    doc_path = base_path / DOC_API_INDEX_PATH
+    with open(doc_path, encoding='utf-8') as f:
         api_index_lines = f.readlines()
 
     doc_api_lines = [line.strip() for line in api_index_lines if line.startswith(API_INDEX_PREFIX)]
@@ -33,20 +34,18 @@ def main(base_path):
 
     ret = 0
     if wmflib_modules - doc_api_modules:
-        print('wmflib modules that are not listed in {doc}: {modules}'.format(
-            doc=DOC_API_INDEX_PATH, modules=wmflib_modules - doc_api_modules))
+        print(f'wmflib modules that are not listed in {DOC_API_INDEX_PATH}: {wmflib_modules - doc_api_modules}')
         ret += 1
     if doc_api_modules - wmflib_modules:
-        print('Documented modules in {doc} that are missing in wmflib: {modules}'.format(
-            doc=DOC_API_INDEX_PATH, modules=doc_api_modules - wmflib_modules))
+        print(f'Documented modules in {DOC_API_INDEX_PATH} that are missing in wmflib: '
+              f'{doc_api_modules - wmflib_modules}')
         ret += 1
 
-    doc_api_files = ['wmflib.{name}.rst'.format(name=name) for name in doc_api_modules]
+    doc_api_files = [f'wmflib.{name}.rst' for name in doc_api_modules]
     missing_doc_api_files = [file for file in doc_api_files
-                             if not os.path.isfile(os.path.join(DOC_API_BASE_PATH, file))]
+                             if not (DOC_API_BASE_PATH / file).is_file]
     if missing_doc_api_files:
-        print('Missing documentation files in {doc}: {files}'.format(
-            doc=DOC_API_BASE_PATH, files=missing_doc_api_files))
+        print(f'Missing documentation files in {DOC_API_BASE_PATH}: {missing_doc_api_files}')
         ret += 1
 
     if ret == 0:
@@ -56,9 +55,9 @@ def main(base_path):
 
 
 if __name__ == '__main__':
-    parser = argparse.ArgumentParser(  # pylint: disable=invalid-name
+    parser = argparse.ArgumentParser(
         description='Check that all wmflib modules are documented')
     parser.add_argument('base_path', help='Path to the root of the wmflib repository')
-    args = parser.parse_args()  # pylint: disable=invalid-name
+    args = parser.parse_args()
 
     sys.exit(main(args.base_path))